+ Added fibonacci number generator recursive task spawning test. + Added cldthread_smart_create() for taking advantage of lazy evaluation in CIEL framework (makes recursive fib effectively iterative). + Added function for querying the master for information about a particular reference. - Fixed bug in CIEL by making implementing a SWReferenceInfoJSONEncoder that doesn't choke when attempting to JSON encode a Task() objects in consumers list. * ReferenceInfo call now returns a future reference if a concrete reference isn't available (really helpful when implementing cldthread_smart_create()). - Fixed bug in reading from wrong cJSON object when deserializing netlocs. * Tweaked cJSON to allow it to read/store integers up to intmax_t (useful when computing Fib(100) :P ).

+++ src/cldthread/libcloudthreads/cldvalue.c
@@ -159,7 +159,7 @@ cldvalue *cldvalue_deserialize( cJSON *json ){
 
             default:
 
-                result->value.data = (void *)cJSON_GetObjectItem(json,"value")->valueint;
+                result->value.data = (void *)(int)cJSON_GetObjectItem(json,"value")->valueint;
                 break;
 
         }
